Factored out lidar_viewer functions into separate file. Added logging.
